intelligence/embeddings/requirements.txt
google-cloud-bigquery>=3.13.0
openai>=1.3.7
google-cloud-aiplatform>=1.38.0

